Porting squeezenet
